Added router for blog entries, it works
